requirements.txt
celery==4.1.1
Django==2.0.6
djangorestframework==3.8.2
requests==2.18.4
# Se hizo downgrade por un problema desconocido...
# pandas==0.24.0
pandas==0.22.0
mysqlclient==1.3.13
SQLAlchemy==1.2.8
redis==2.10.6
uWSGI==2.0.17
django-crispy-forms==1.7.2
SQLAlchemy-Utils==0.33.3
raven==6.9.0
openpyxl==2.5.5
flower==0.9.2
tornado>=4.2.0,<6.0.0
kombu==4.2.1
flask==1.0.2